Ok the board is solid
got the port and got the board to genereate the mines in more of random way and  it looks cool. what else well the next step is to make a second boat to navigate with to the end.
also need to make it depend on whether the player has the minesweeper boat