New Game going
so the player needs to escort the 2 boats to the other side of the map and they have their own button sets
so the idea is that the minesweeper can mark paths to be safe so the boats dont take random damage.
there are 2 new classes that are minesweeper and battleship and they derive from the playercontrol class that is now abstract.